+++ mods/BrewLAN/hook/lua/seraphimweapons.lua
@@ -1,0 +1,8 @@
+
+BrewLANUltraChonkChromBeamGenerator = Class(SDFUltraChromaticBeamGenerator) {
+    BeamType = CollisionBeamFile.ChonkChromBeamGenCollisionBeam,
+}
+
+BrewLANTracerChromBeamGenerator = Class(SDFUltraChromaticBeamGenerator) {
+    BeamType = CollisionBeamFile.TracerChromBeamGenCollisionBeam,
+}